did cosmetic changes related to the save report feature.

+++ public/javascripts/adv_reports/adv_reports.js
@@ -182,9 +182,13 @@ Helpkit.reports_util = {
   bindEventsForReportFilter: function(){
     /** event binds related to saved report starts here**/
     jQuery(document).on('click','.filter-event',function(){
-      var _id = jQuery(this).attr('data-id');
+      var _id      = jQuery(this).attr('data-id'),
+          _jq_this = jQuery(this);
+
       Helpkit.reports_util.updateReportFilters(_id);
-      jQuery("#active_filter").text(jQuery(this).text());
+      jQuery("#active_filter").text(_jq_this.text());
+      jQuery(".selected-filter").removeClass('selected-filter');
+      _jq_this.addClass('selected-filter');
       jQuery("#delete-filter-link").removeClass('hide').attr('data-id',_id);
       jQuery("#save-filter-link").addClass('hide');
     });
@@ -202,10 +206,12 @@ Helpkit.reports_util = {
       jQuery("#report-dialog-select-cancel").click();
     });
     jQuery(document).on('click',"#delete-filter-link",function(){
-      Helpkit.reports_util.deleteReportFilter(jQuery(this).attr('data-id'));
-      jQuery("#active_filter").text("Unsaved Report");
-      jQuery("#save-filter-link").removeClass('hide');
-      jQuery("#delete-filter-link").addClass('hide');
+      if(window.confirm('Are you sure you want to delete this saved report ?')){
+        Helpkit.reports_util.deleteReportFilter(jQuery(this).attr('data-id'));
+        jQuery("#active_filter").text("Unsaved Report");
+        jQuery("#save-filter-link").removeClass('hide');
+        jQuery("#delete-filter-link").addClass('hide');
+      }
     });
     /** event binds related to saved report ends here**/
   },
